Moved Product settings sections and rows in a separate file
